Add a 'branch_only' parameter for /module
When passing this parameter, it tries to detect the branch from the
module as well and apply it using the -e parameter to r10k. For modules
that are branch-linked with the main repository, this will trigger an
update *only* to the respective environment, rather than all
environments.

* For pushes into a branch of the module corresponding to an
environment, this merely saves processing time.
* For pushes to the default branch (of the module) which would affect
all environments, this will block updates to all environments except
for the default one (ie. production). This helps prevent unsolicited
updates of an environment that could break it, until the environment
itself is updated by other means (eg. by pushing into the respective
branch of the source repository or module).