Use provided r10k command_path
Prior to this, the Config struct had a setting under the R10k struct
called CommandPath that could be set in the config file, but was
ignored as both places that would use it were hard coded to instead use
the string `r10k`. This resulted in the application looking in the path
for the r10k binary. A default is set in config.go also, but that seems
to have also been ignored.

This fixes #152