add mechanism for blocking branches
Added the following to allow for branches to be blocked from deployment
by Webhook/Puppet admin:

* Added new String Slice (`[]string`) called `BlockedBranches`
("blocked_branches" in YAML file). This is an array/slice of branch
names to block from deployment
* Added logic to the environment and module api endpoints that cancel
the request if a blocked branch is detected. A `403 Forbidden` will be
returned as a response and an Error will be logged.
* Updated README with new config option